CLI "add <vg>" now is able to spawn a specific thread.
The thread is actually not doing anything now, because the main thread
is still using the default one.
A pool of VGs is introduced: it is a list of VG oriented structures
that couple together VG, threads and their own request queue.

So far, stopping the thread is not possible, too.

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>