Default worker thread using evenfd as notification system
The standard notification system for threads (cond_wait) cannot
be used for threads that potentially need to wait for other
events, such as network events.

The default thread in the network scenario is supposed to handle TCP
requests, as well.

Eventfd could became the default mechanism for all the threads once
proven reliable and efficient

Signed-off-by: Germano Percossi <germano.percossi@citrix.com>